+++ packages/platforms/react-native/tests/auto-instrumentation/network-request-plugin.test.ts
@@ -1,11 +1,12 @@
-import { spanContextEquals } from '@bugsnag/core-performance'
-import { MockSpanFactory, createConfiguration, createTestClient } from '@bugsnag/js-performance-test-utilities'
+import { DefaultSpanContextStorage, type SpanContextStorage, spanContextEquals } from '@bugsnag/core-performance'
+import { ControllableBackgroundingListener, MockSpanFactory, createConfiguration, createTestClient } from '@bugsnag/js-performance-test-utilities'
 import { type ReactNativeSchema, type ReactNativeConfiguration } from '../../lib/config'
 import { RequestTracker, type RequestStartCallback } from '@bugsnag/request-tracker-performance'
 import { NetworkRequestPlugin } from '../../lib/auto-instrumentation/network-request-plugin'
 
 const ENDPOINT = 'http://traces.endpoint'
 const TEST_URL = 'http://test-url.com/'
+const SAME_ORIGIN_TEST_URL = 'http://localhost/my-api'
 
 class MockRequestTracker extends RequestTracker {
   onStart = jest.fn((startCallback: RequestStartCallback) => {
@@ -13,49 +14,57 @@ class MockRequestTracker extends RequestTracker {
   })
 }
 
+const createConfig = (overrides: Partial<ReactNativeConfiguration> = {}) => {
+  return createConfiguration<ReactNativeConfiguration>({
+    endpoint: ENDPOINT,
+    autoInstrumentNetworkRequests: true,
+    tracePropagationUrls: [],
+    ...overrides
+  })
+}
+
 describe('network span plugin', () => {
   let xhrTracker: MockRequestTracker
   let spanFactory: MockSpanFactory<ReactNativeConfiguration>
+  let spanContextStorage: SpanContextStorage
 
   beforeEach(() => {
     xhrTracker = new MockRequestTracker()
     spanFactory = new MockSpanFactory()
+    spanContextStorage = new DefaultSpanContextStorage(new ControllableBackgroundingListener())
   })
 
   it('tracks requests when autoInstrumentNetworkRequests = true', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrTracker)
     expect(xhrTracker.onStart).not.toHaveBeenCalled()
 
-    plugin.configure(createConfiguration<ReactNativeConfiguration>({
-      endpoint: ENDPOINT,
-      autoInstrumentNetworkRequests: true
-    }))
+    plugin.configure(createConfig())
 
     expect(xhrTracker.onStart).toHaveBeenCalled()
   })
 
   it('starts a span on request start', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrTracker)
 
-    plugin.configure(createConfiguration<ReactNativeConfiguration>({
-      endpoint: ENDPOINT,
-      autoInstrumentNetworkRequests: true
-    }))
+    plugin.configure(createConfig())
 
-    xhrTracker.start({ type: 'xmlhttprequest', method: 'POST', url: TEST_URL, startTime: 2 })
-    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/POST', { startTime: 2, makeCurrentContext: false })
+    const res = xhrTracker.start({ type: 'xmlhttprequest', method: 'POST', url: TEST_URL, startTime: 2 })
+    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP/POST]', { startTime: 2, makeCurrentContext: false })
+    expect(res.extraRequestHeaders).toEqual([])
+
+    // currently traceparent headers are not added to any requests by default
+    const res2 = xhrTracker.start({ type: 'xmlhttprequest', method: 'POST', url: SAME_ORIGIN_TEST_URL, startTime: 2 })
+    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP/POST]', { startTime: 2, makeCurrentContext: false })
+    expect(res2.extraRequestHeaders).toEqual([])
   })
 
   it('ends a span on request end', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrTracker)
 
-    plugin.configure(createConfiguration<ReactNativeConfiguration>({
-      endpoint: ENDPOINT,
-      autoInstrumentNetworkRequests: true
-    }))
+    plugin.configure(createConfig())
 
     const { onRequestEnd: endRequest } = xhrTracker.start({ type: 'xmlhttprequest', method: 'GET', url: TEST_URL, startTime: 1 })
-    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/GET', { startTime: 1, makeCurrentContext: false })
+    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP/GET]', { startTime: 1, makeCurrentContext: false })
     expect(spanFactory.endSpan).not.toHaveBeenCalled()
 
     endRequest({ status: 200, endTime: 2, state: 'success' })
@@ -63,7 +72,7 @@ describe('network span plugin', () => {
     expect(spanFactory.createdSpans.length).toEqual(1)
 
     const span = spanFactory.createdSpans[0]
-    expect(span.name).toEqual('[HTTP]/GET')
+    expect(span.name).toEqual('[HTTP/GET]')
     expect(span.startTime).toEqual(1)
     expect(span.endTime).toEqual(2)
     expect(span).toHaveAttribute('bugsnag.span.category', 'network')
@@ -73,36 +82,27 @@ describe('network span plugin', () => {
   })
 
   it('does not track requests when autoInstrumentNetworkRequests = false', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrTracker)
 
-    plugin.configure(createConfiguration<ReactNativeConfiguration>({
-      endpoint: ENDPOINT,
-      autoInstrumentNetworkRequests: false
-    }))
+    plugin.configure(createConfig({ autoInstrumentNetworkRequests: false }))
 
     expect(xhrTracker.onStart).not.toHaveBeenCalled()
   })
 
   it('does not track requests to the configured traces endpoint', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrTracker)
 
-    plugin.configure(createConfiguration<ReactNativeConfiguration>({
-      endpoint: ENDPOINT,
-      autoInstrumentNetworkRequests: true
-    }))
+    plugin.configure(createConfig())
 
     xhrTracker.start({ type: 'xmlhttprequest', method: 'GET', url: `${ENDPOINT}`, startTime: 1 })
     expect(spanFactory.startSpan).not.toHaveBeenCalled()
   })
 
   it('does not track requests to the NetInfo reachability URL', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrTracker)
     const NET_INFO_REACHABILITY_URL = 'https://clients3.google.com/generate_204?_=1701691583660'
 
-    plugin.configure(createConfiguration<ReactNativeConfiguration>({
-      endpoint: ENDPOINT,
-      autoInstrumentNetworkRequests: true
-    }))
+    plugin.configure(createConfig())
 
     xhrTracker.start({ type: 'xmlhttprequest', method: 'GET', url: NET_INFO_REACHABILITY_URL, startTime: 1 })
     expect(spanFactory.startSpan).not.toHaveBeenCalled()
@@ -118,12 +118,9 @@ describe('network span plugin', () => {
   ]
 
   it.each(expectedProtocols)('tracks requests over all expected protocols', (url) => {
-    const plugin = new NetworkRequestPlugin(spanFactory, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrTracker)
 
-    plugin.configure(createConfiguration<ReactNativeConfiguration>({
-      endpoint: ENDPOINT,
-      autoInstrumentNetworkRequests: true
-    }))
+    plugin.configure(createConfig())
 
     xhrTracker.start({ type: 'xmlhttprequest', method: 'GET', url, startTime: 1 })
     expect(spanFactory.startSpan).toHaveBeenCalled()
@@ -142,49 +139,42 @@ describe('network span plugin', () => {
   ]
 
   it.each(unexpectedProtocols)('does not track requests over unexpected protocols', (url) => {
-    const plugin = new NetworkRequestPlugin(spanFactory, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrTracker)
 
-    plugin.configure(createConfiguration<ReactNativeConfiguration>({
-      endpoint: ENDPOINT,
-      autoInstrumentNetworkRequests: true
-    }))
+    plugin.configure(createConfig())
 
     xhrTracker.start({ type: 'xmlhttprequest', method: 'GET', url, startTime: 1 })
     expect(spanFactory.startSpan).not.toHaveBeenCalled()
   })
 
   it('discards the span if the status is 0', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrTracker)
 
-    plugin.configure(createConfiguration<ReactNativeConfiguration>({
-      endpoint: ENDPOINT,
-      autoInstrumentNetworkRequests: true
-    }))
+    plugin.configure(createConfig())
 
     const { onRequestEnd: endRequest } = xhrTracker.start({ type: 'xmlhttprequest', method: 'GET', url: TEST_URL, startTime: 1 })
-    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/GET', { startTime: 1, makeCurrentContext: false })
+    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP/GET]', { startTime: 1, makeCurrentContext: false })
 
     endRequest({ endTime: 2, state: 'error' })
     expect(spanFactory.endSpan).not.toHaveBeenCalled()
   })
 
   it('discards the span if there is an error', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrTracker)
 
-    plugin.configure(createConfiguration<ReactNativeConfiguration>({
-      endpoint: ENDPOINT,
-      autoInstrumentNetworkRequests: true
-    }))
+    plugin.configure(createConfig())
 
     const { onRequestEnd: endRequest } = xhrTracker.start({ type: 'xmlhttprequest', method: 'GET', url: TEST_URL, startTime: 1 })
-    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/GET', { startTime: 1, makeCurrentContext: false })
+    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP/GET]', { startTime: 1, makeCurrentContext: false })
 
     endRequest({ state: 'error', error: new Error('woopsy'), endTime: 2 })
     expect(spanFactory.endSpan).not.toHaveBeenCalled()
   })
 
   it('does not push network spans to the context stack', () => {
-    const client = createTestClient<ReactNativeSchema, ReactNativeConfiguration>({ plugins: (spanFactory) => [new NetworkRequestPlugin(spanFactory, xhrTracker)] })
+    const client = createTestClient<ReactNativeSchema, ReactNativeConfiguration>(
+      { plugins: (spanFactory, spanContextStorage) => [new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrTracker)] }
+    )
     const rootSpan = client.startSpan('root span')
 
     xhrTracker.start({ type: 'xmlhttprequest', method: 'POST', url: TEST_URL, startTime: 2 })
@@ -193,10 +183,8 @@ describe('network span plugin', () => {
   })
 
   it('prevents creating a span when networkRequestCallback returns null', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, xhrTracker)
-    plugin.configure(createConfiguration<ReactNativeConfiguration>({
-      endpoint: ENDPOINT,
-      autoInstrumentNetworkRequests: true,
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrTracker)
+    plugin.configure(createConfig({
       networkRequestCallback: (networkRequestInfo) => networkRequestInfo.url === 'no-delivery' ? null : networkRequestInfo
     }))
 
@@ -204,14 +192,12 @@ describe('network span plugin', () => {
     expect(spanFactory.startSpan).not.toHaveBeenCalled()
 
     xhrTracker.start({ type: 'xmlhttprequest', method: 'GET', url: TEST_URL, startTime: 1 })
-    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/GET', { startTime: 1, makeCurrentContext: false })
+    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP/GET]', { startTime: 1, makeCurrentContext: false })
   })
 
   it('uses a modified url from networkRequestCallback', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, xhrTracker)
-    plugin.configure(createConfiguration<ReactNativeConfiguration>({
-      endpoint: ENDPOINT,
-      autoInstrumentNetworkRequests: true,
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrTracker)
+    plugin.configure(createConfig({
       networkRequestCallback: (networkRequestInfo) => ({
         type: networkRequestInfo.type,
         url: 'modified-url'
@@ -219,19 +205,70 @@ describe('network span plugin', () => {
     }))
 
     const { onRequestEnd: endRequest } = xhrTracker.start({ type: 'fetch', method: 'GET', url: TEST_URL, startTime: 1 })
-    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/GET', { startTime: 1, makeCurrentContext: false })
+    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP/GET]', { startTime: 1, makeCurrentContext: false })
 
     endRequest({ status: 200, endTime: 2, state: 'success' })
     expect(spanFactory.endSpan).toHaveBeenCalled()
     expect(spanFactory.createdSpans.length).toEqual(1)
 
     const span = spanFactory.createdSpans[0]
-    expect(span.name).toEqual('[HTTP]/GET')
+    expect(span.name).toEqual('[HTTP/GET]')
     expect(span.startTime).toEqual(1)
     expect(span.endTime).toEqual(2)
     expect(span).toHaveAttribute('bugsnag.span.category', 'network')
     expect(span).toHaveAttribute('http.url', 'modified-url')
     expect(span).toHaveAttribute('http.method', 'GET')
     expect(span).toHaveAttribute('http.status_code', 200)
+  })
+
+  it('adds trace propagation headers when url matches tracePropagationUrls (string)', () => {
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrTracker)
+    plugin.configure(createConfig({
+      tracePropagationUrls: [TEST_URL]
+    }))
+
+    const res = xhrTracker.start({ type: 'fetch', method: 'GET', url: TEST_URL, startTime: 1 })
+    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP/GET]', { startTime: 1, makeCurrentContext: false })
+    expect(res.extraRequestHeaders).toEqual([
+      { traceparent: '00-a random 128 bit string-a random 64 bit string-01' }
+    ])
+  })
+
+  it('adds trace propagation headers when url matches tracePropagationUrls (RegExp)', () => {
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrTracker)
+    plugin.configure(createConfig({
+      tracePropagationUrls: [
+        /^(http(s)?(:\/\/))?(www\.)?test-url\.com(\/.*)?$/
+      ]
+    }))
+
+    const res = xhrTracker.start({ type: 'fetch', method: 'GET', url: TEST_URL, startTime: 1 })
+    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP/GET]', { startTime: 1, makeCurrentContext: false })
+    expect(res.extraRequestHeaders).toEqual([
+      { traceparent: '00-a random 128 bit string-a random 64 bit string-01' }
+    ])
+  })
+
+  it('does not add trace propagation headers when tracePropagationUrls is empty', () => {
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrTracker)
+    plugin.configure(createConfig())
+
+    const res = xhrTracker.start({ type: 'fetch', method: 'GET', url: SAME_ORIGIN_TEST_URL, startTime: 1 })
+    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP/GET]', { startTime: 1, makeCurrentContext: false })
+    expect(res.extraRequestHeaders).toEqual([])
+  })
+
+  it('does not add trace propagation headers when url does not match tracePropagationUrls', () => {
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, xhrTracker)
+    plugin.configure(createConfig({
+      tracePropagationUrls: [
+        'http://noheader.com/',
+        /^(http(s)?(:\/\/))?(www\.)?noheader\.com(\/.*)?$/
+      ]
+    }))
+
+    const res = xhrTracker.start({ type: 'fetch', method: 'GET', url: SAME_ORIGIN_TEST_URL, startTime: 1 })
+    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP/GET]', { startTime: 1, makeCurrentContext: false })
+    expect(res.extraRequestHeaders).toEqual([])
   })
 })