+++ packages/request-tracker/lib/network-request-callback.ts
@@ -1,9 +1,5 @@
 export interface NetworkRequestInfo {
   url: string | null
-  /**
-   * Experimental. Whether to propagate trace context by adding a `traceparent` header to the request.
-   */
-  propagateTraceContext?: boolean
 }
 
 export type NetworkRequestCallback <T extends NetworkRequestInfo> = (networkRequestInfo: T) => T | null